Moved Panorama to a module folder
