+++ lib/src/bot.dart
@@ -10,7 +10,7 @@ import 'package:weather/src/modules/chat/chat.dart';
 import 'package:weather/src/modules/chat/chat_manager.dart';
 import 'package:weather/src/modules/user/user_manager.dart';
 import 'package:weather/src/modules/weather_manager.dart';
-import 'package:weather/src/modules/panorama_manager.dart';
+import 'package:weather/src/modules/panorama/panorama_manager.dart';
 import 'package:weather/src/modules/dadjokes/dadjokes_manager.dart';
 import 'package:weather/src/modules/reputation_manager.dart';
 import 'package:weather/src/modules/youtube_manager.dart';
