+++ lib/src/modules/panorama/panorama.dart
