+++ lib/src/modules/panorama/panorama_manager.dart
@@ -1,11 +1,11 @@
-import 'package:weather/src/core/panorama.dart';
 import 'package:weather/src/globals/chat_platform.dart';
 import 'package:weather/src/globals/message_event.dart';
 import 'package:weather/src/injector/injection.dart';
 import 'package:weather/src/platform/platform.dart';
 import 'package:weather/src/modules/chat/chat.dart';
-import 'package:weather/src/modules/utils.dart';
 import 'package:weather/src/utils/logger.dart';
+import 'panorama.dart';
+import '../utils.dart';
 
 class PanoramaManager {
   final Platform platform;
